+++ komikaan/Interfaces/ITravelAdviceHandler.cs
@@ -1,0 +1,9 @@
+﻿using komikaan.Models;
+
+namespace komikaan.Interfaces
+{
+    public interface ITravelAdviceHandler
+    {
+        Task<JourneyResult> GetTravelExpectationAsync(string fromStop, string toStop);
+    }
+}