+++ komikaan/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.7.1] - 17-03-2024
+### Added
+- OpenTelemetry support
+### Changed 
+- Move logic out of controllers
+
+[1.7.0] - 17-03-2024
+### Added
+- Support for stops along the trip
+### Changed
+- AffectedStops of disruptions are now SimpleStops's
+- NS - If there are no affectedStops defined, unset it and assume it applies to everything
+
 [1.6.0] - 08-03-2024
 ### Added
 - Stop Manager to merge related stops of different suppliers